+++ backend/utils/getIP.js
@@ -1,0 +1,14 @@
+function getIPAddress() {
+    const interfaces = require('os').networkInterfaces();
+    for (const interfaceName in interfaces) {
+        const iface = interfaces[interfaceName];
+        for (const alias of iface) {
+            if (alias.family === 'IPv4' && !alias.internal) {
+                return alias.address;
+            }
+        }
+    }
+    return '127.0.0.1'; // Fallback to localhost
+}
+
+module.exports = getIPAddress;
