+++ frontend/src/pages/Views/desktop/Profile.js
@@ -22,7 +22,7 @@ export default function Profile() {
   const { user, setUser } = useAuth();
 
   const serverUrl =
-    process.env.REACT_APP_BACKEND_URL || "http://192.168.1.8:5000";
+    process.env.REACT_APP_BACKEND_URL || "http://192.168.8.168:5000";
 
   useEffect(() => {
     const fetchUserData = async () => {
